Move GameCard to common dir
