+++ frontend/app/ui/pong/GameCard.tsx
@@ -4,7 +4,7 @@ import { Avatar } from "@/app/ui/user/avatar";
 import { Button } from "@/components/ui/button";
 import { Card } from "@/components/ui/card";
 import { TooltipProvider } from "@radix-ui/react-tooltip";
-import { PublicUserEntity } from "../lib/dtos";
+import { PublicUserEntity } from "../../lib/dtos";
 
 function UserCard({ user }: { user: PublicUserEntity }) {
   return (
